[Feature] DeepSeek-V4 for Ascend950 (#9757)
###  What this PR does / why we need it?
  
This PR introduces DeepSeek-V4 model support for the Ascend A5
(Ascend950) device. It adds several key capabilities:

  1. W4A8 MXFP (Microscaling) Quantization
- Introduces a new QuantType.W4A8MXFP quantization type with dedicated
linear and MoE methods (vllm_ascend/quantization/methods/w4a8_mxfp4.py,
  vllm_ascend/quantization/methods/fp8.py).
- Adds AscendFp8Config (vllm_ascend/quantization/fp8_config.py) to serve
as the quantization configuration entry point for DeepSeek-V4 FP8 models
on
Ascend, integrating with both the "fp8" and "deepseek_v4_fp8"
quantization method names.
- Implements A5-specific MXFP quantized matmul and grouped
matmul-swish-glu operators in the device adaptor layer
(vllm_ascend/device/device_op.py),
leveraging NPU hardware instructions for float4_e2m1fn_x2 weight formats
with float8_e8m0fnu microscaling.
  
  2. Ascend A5 Attention Kernel Optimizations
- Major overhaul of the direct sparse attention (DSA) v1 kernel
(vllm_ascend/attention/dsa_v1.py, +454/-188) to support A5-specific
block sizes, scheduling
  metadata, and KV cache interfaces.
- A5-specific compile-time/scheduling adaptations for decode and prefill
attention paths.
  
  3. DeepSeek-V4 Model Architecture Support
- Extends DeepseekV4MoE and related modules with swiglu_limit support
and A5-specific quant_config handling for compressor projections
  (vllm_ascend/models/deepseek_v4.py).
- Adjusted compressor block_size from 32 to 16 for A5 device
compatibility.
  
  4. FusedMoE Enhancements
- Updated the FusedMoE pipeline (contracts, stage params,
prepare/finalize) to pass through W4A8 MXFP quantization parameters
(activation quantization
  dtype, weight quantization dtype, scale dtypes, BF16 usage flag).
  
### Why we need it: DeepSeek-V4 is a new generation model that cannot
run on Ascend A5 without these adaptations. The A5 NPU has specific
hardware requirements
for weight layout, quantization format, and kernel block sizes that
differ from previous Ascend generations. This PR is the first step to
enable
  DeepSeek-V4 inference on Ascend950 hardware.

  ---
  Does this PR introduce any user-facing change?
  
  Yes:
- DeepSeek-V4 can now be served on Ascend A5 hardware with W4A8 MXFP
quantization by specifying --quantization fp8 (or deepseek_v4_fp8) in
the vLLM serve
  command.
- A new quantization type W4A8MXFP is available in the Ascend
quantization registry.
- No API or CLI changes; the feature is activated automatically when the
Ascend device type is A5 and the model config specifies a supported
quantization
  format.

  ---
 ### How was this patch tested?
  
- Unit tests: updated tests/ut/ops/test_moe_runtime_args.py to cover the
new W4A8 MXFP quantization path in MoE runtime argument construction.
- The PR is labeled module:tests, module:ops, module:core, and
module:quantization, indicating test coverage across all modified
subsystems.
- End-to-end validation was performed by running DeepSeek-V4 inference
on Ascend A5 hardware with W4A8 MXFP quantized weights.

  ---


- vLLM version: v0.20.2
- vLLM main:
https://github.com/vllm-project/vllm/commit/9090368b650896bf5fc990c921df7eb4c20355a5

---------

+++ vllm_ascend/ops/fused_moe/moe_comm_method.py
@@ -81,7 +81,7 @@ class FusedExpertsResult:
     # For dynamic_eplb
     group_list_type: int = 1
     expert_tokens: torch.Tensor | None = None
-    swiglu_limit: int = 0
+    swiglu_limit: float = 0.0
 
 
 class MoECommMethod(ABC):
