+++ vllm_ascend/ops/fused_moe/moe_stage_contracts.py
@@ -68,7 +68,7 @@ class MoEFusedExpertsInput:
     activation: str = "silu"
     need_trans: bool = False
     dynamic_eplb: bool = False
-    swiglu_limit: int = 0
+    swiglu_limit: float = 0.0
 
 
 @dataclass(frozen=True, slots=True)
@@ -140,7 +140,7 @@ class MoEMlpComputeInput:
     activation: str = "silu"
     need_trans: bool = False
     dynamic_eplb: bool = False
-    swiglu_limit: int = 0
+    swiglu_limit: float = 0.0
 
 
 __all__ = [
